Use the Instruction Address as Temporary Scalar Index
This replaces the "unique temporary index per block" approach.

The unique temporary index per block is almost always 0 because
when lifting the instruction from Capstone to Falcon IL, each
lifted instruction results in a new CFG.

When merging the CFGs the temporary scalars are not adjusted,
therefore we end up with a lot of duplicated `temp_0.0` scalars
(which may have different bitness!).

This may cause problems when the CFG is manipulated and later
transformed into SSA.