Graph: Remove head field (#69)
None of the functions in `Graph` really makes use of the head field.

Additionally, the head field may easily become out of sync with the
CFG entry resulting in subtile bugs, such as removing the wrong
unreachable vertices. This can be avoided by removing the head field
altogether and instead require it as function argument.